Make ukernels assert their validation, remove `iree_uk_status_t`. (#11900)
Benefits:
- Now we actually see error messages in e2e tests when ukernels generate
an error. Before, error statuses were silently eaten. This fixes #11862
(in a somewhat radical way).
 - Net removal of > 100 lines.

The main difficulty is that ukernels can't use the standard library, and
one can't practically implement assertions without that. This is solved
by moving the assert failure function out of the ukernels library, into
its own, which callers may either link against or reimplement in a way
that works for them.

Fixes #11862 .