Make sure the selected workgroup size divides the problem size. (#11907)
Without either pre-padding or masking not having the tile sizes divide the problem size for all tiling levels will result in un-vectorized code that has the side-effect of stack allocation. Even though this stack is bounded it is within the inner most loop and dynamic which will result in stack overflow. The matmul path is supposed to use pad + hoist, but that doesnt seem to kick in for non x86 or non f32 paths.
This workaround addresses this in the short-term, avoid stack allocations. A proper fix will need to either deprecate this pipeline completely, use vector masking or use pre-padding/data tiling.

Issue #11880