Add a `--iree-preprocessing-pass-pipeline` to allow user control on preprocessing passes before IREE compilation. (#11986)
This PR adds the --iree-preprocessing-pass-pipeline that allows users to run a sequence of passes (using the MLIR -pass-pipeline syntax) to control the sequnece of passes that do preprocessing of the program after conversion from Input dialects (like MHLO/TOSA), before running any of the core IREE compilation pipelines. This allows user control on using things like Winograd/Im2Col transformation etc. that is geared towards specific use cases. This could also be used as a experimental/prototyping space. If there are transformations that apply to all backends and all models, they could be moved out of preprocessing into the core IREE compilation pipelines.

This also removes the need to add some global options to iree-compile, reducing the number of -iree-flow-* options. More will be removed in the future.