Allow contolling the default workgroup tile size for unpack op (#12134)
The value 16 here didn't work well. It's in destination space, so for
e.g. a 8x8 tile size, the default 16x16 means unpacking only a tiny
2x2x8x8 tile, not enough to reach high performance in a ukernel. Just
the generic `defaultWorkgroupTileSize` used for other ops, with its
default value 64, gives me a large speedup by default and allows me to
tune this more easily along with the rest.