Get the right `pack`/`unpack` ukernel ARM64 tile functions (#12141)
- `pack` ukernel:
- Add missing tile function for the case of 8x8 tiles of 32-bit
elements. This corresponds to accumulator tiles, and I had overlooked
that as I had focused on the LHS/RHS tile cases.
- `unpack` ukernel:
- I had blithely copied my earlier `pack` tile functions into `unpack`
tile functions. This means that not only I implemented a lot of `unpack`
cases that are unused in practice because they correspond to LHS/RHS
tiles that are only ever packed, not unpack, but as noted above the
`pack` code was missing the tile for the accumulator case and so I also
forgot that in `unpack`, even though that's the only thing that matters
in `unpack`!

Deleting for now all the code that isn't immediately used. It's still
there in git history if we need it.