Defaulting verbose debug info in LLVM CPU IR to off. (#12155)
Adds the `--iree-codegen-llvm-verbose-debug-info=` flag to control
whether we emit per arg/field debug info. This debug info is useful if
using a debugger and stepping into codegen emitted dispatch routines but
MLIR is currently not very usable with so much additional debug
information. All the extra debug info disables optimizations and
obscures program flow (e.g. CSE breaks if DI ops are present) and leads
to unreadable IR dumps due to inline attribute printing including both
the per-op debug info as well as the debug info for the entire parent
scope on each op.

Since the use case for this is rather limited today we'll default off
for now and in the future perhaps turn it on via higher-level
`-gN`-style flags.

Fixes #12114.