Replace `zve32x` with `zve32f` in CI configurations (#12185)
The zve32x feature includes support for integer vectors but not fp vectors. When we send fp vector ops to the RISC-V backend with only zve32x support, they should be scalarized. Unfortunately, this is currently not implemented in the backend (see llvm/llvm-project#60463). In the meantime, we are replacing 'zve32x' with 'zve32f', which also includes support for fp vector operations.