Adding `--iree-hal-substitute-executable-*=` flags. This allows for specifying one or more `executable_name=file.xxx` pairs that each replace a `hal.executable` op with the `executable_name` with the contents of `file.xxx`. .mlir/.mlirbc files are loaded and a `hal.executable` with the matching name is used as a replacement while any other file type will cause the original executable to be externalized and linked with the specified file (.ptx/.spv/etc).
The additional `--iree-hal-substitute-executable-*-from=` flags allow
for scanning a directory for executables by name to build the
substitution mapping. Only files in the executable_name or
module_executable_name form will be substituted but we can extend this
in the future to support variant naming.

Because of phase ordering constraints around where codegen is able to
mutate host-related code such as workgroup count calculations there are
two flag sets:
  `--iree-hal-substitute-executable-source=name=file.xxx`
  `--iree-hal-substitute-executable-sources-from=path/`
 and
  `--iree-hal-substitute-executable-object=name=file.xxx`
  `--iree-hal-substitute-executable-objects-from=path/`
Sources are substituted immediately prior to benchmark generation and
when substituting target objects (.ptx, .spv, etc) require that it's ok
to skip codegen (workgroup count calculation is not dependent on root
op detection, etc). Objects are substituted immediately after codegen
for use in cases where codegen is to generate the host code. There are
uses for both depending on what the input IR is and what the developer
wants to modify (host code, device code, or both).

The primary developer workflows this covers are:
1. dump executable sources via `--iree-hal-dump-executable-sources-to=`
   and modify them, potentially running any number of iree-opt passes,
   before linking them back in to the original program they came from
2. author custom implementations ala the custom_dispatch sample in target
   toolchains (.cu -> .ptx, .glsl -> .spv, etc) and use those in full
   programs without needing to modify the compiler
3. do either of the above and use the substituted executable for
   microbenchmarking via `--iree-hal-dump-executable-benchmarks-to=` (so
   one can easily microbenchmark handwritten kernels)

Progress on #12222 (rest is orthogonal).