[GlobalOptimization] Add a pass to convert convolutions to channels last
This pass allows converting all NCHW-like convolution ops to NHWC. The
better way to implement the functionality in this pass is with encodings
the same way that matmul tiling is handled today, however there are
a couple differences with convolution that motivate this pass right now.

1) The immediate input to a convolution is typically a tensor.pad, so
   without basic propagation through that pad, we can end up with two
   dispatches for the pack and pad today. This is essentially a bug in
   fusion + codegen, however handling it in this pass allows staging
   the necessary work.

2) Pack and unpack lack proper handling in some backends, and the
   default state for this pass is to just use it for transposing
   convolutions.

Currently this includes some baked in generalization patterns and does
not support padding. Tile size selection currently is pass-wide, but
there is limited attribute control to enable fully transposing. Further
generalizations should aim to write this pass by allowing per-op tile
size control.

The plan is to add some common useful preprocessing pipelines as
syntactic sugar based on this pass.