Add a bring-your-own-LLVM build script. (#12274)
This wraps the following into one reference flow (that we can run on
CI):

1. Build and install LLVM+tools (llvm, lld, clang).
2. Build and install MLIR standalone against installed LLVM.
3. Build IREE against separately installed LLVM and MLIR.

Supporting a multi-step like this may help division of labor in shops
who have a compiler team focused on building an LLVM toolchain as it
will enable them to build that apart from the MLIR churn. Folks would
then layer a locally built MLIR on top of that and build IREE on top of
the combination.

Builds on #12256.
Progress on #12231.