Handle workflow dispatch in ci-stage determination (#12482)
Otherwise `workflow_dispatch` events will act like postsubmits, posting
on failures and uploading benchmark results. I'm not categorizing
`workflow_dispatch` as "presubmit" here because that implies that it
came from a PR which also may not be the case and is problematic
because dependent actions might reasonably assume that they can query
the PR info.