Fully retire CanonicalizedSequenceOp (#12467)
This revision fully retires the CanonicalizedSequenceOp and makes the
need for canonicalization / tiling_canonicalization / licm / singleIterationLoopPromotion
explicit in the transform IR.
    
This removes automatic blanket pattern applications which is expected to be beneficial for:
1. reducing compile time
2. reducing surprise (explicit being better than implicit)
3. help surface issues similar to #12444

In the process, plumb the isolatedFromAbove through the API to make the
requirement explicit until we are able to remove it in some future (i.e. with better targeted
canonicalization and greedy pattern rewriter that does not require an isolated from above).