[Codegen] Add canonicalization pattern for reshape into interface stores (#12471)
This pattern is similar to the existing one for loads except for stores. This is needed for implicit gemm where a reshape ends up at the bottom of the kernel.This pattern is similar to the existing one for loads except for stores. This is needed for implicit gemm where a reshape ends up at the bottom of the kernel.