For float32 kernels, check for "avx" feature, not "avx2" (#12676)
AVX2 brings support for non-float32 data types, but for float32, AVX is
what we need.

Separately we will also check for FMA support in the ukernel impl, and
AVX2 is often used as a proxy for AVX+FMA but it's technically not the
same even if in practice, CPUs support AVX2 if and only if they support
AVX+FMA. And for those targets that support AVX but not FMA (e.g. Sandy
Bridge), thanks to checking for AVX here we will get the right tile size
for that case and then codegen should be able to save the day.