select aarch64 specific tests on Apple arm64 (#12674)
I was wondering why we didn't catch on Mac/ARM CI that aarch64
regression from llvm-integrate, which we only saw on Android device
tests (#12546). The reason was that our CMake logic only enabled those
tests when CMAKE_SYSTEM_PROCESSOR is "aarch64", but on Apple it is
"arm64" instead. We have some other places in the code that are already
checking for both values, but this place was missing it.

@wangkuiyi I haven't tested this locally as I haven't set up my Mac for
local development.