[TransformStrategies] Enable aligned and partially aligned matmul
Current handling of tile aligned matmuls differs from unaligned in the
way the tensor.pad ops are lowered. When the pad is aligned and would
otherwise fold, we convert to a linalg.copy and don't need to do masked
vectorization. In the future the plan is to unify aligned and unaligned
cases.