[GlobalOptimization] Add raising pattern for type extensions into named matmul ops
A matmul has implicit upcasting semantics of its operands to the
accumulator type. This makes it more canonical to fold upcast ->
linalg.matmul rather than keeping two separate ops to inform
tiling/instruction selection decisions, as well as improves the ability
of other GlobalOptimization/Flow patterns to work with the matmuls.

The reason this only applies to named operations is that this is already
modelled by ElementwiseOpFusion for generics. Named ops were the only
ones missing this kind of simplification.