[Codegen][GPU] Add deferred thread lowering config attribute
This adds a lowering config that defers the concrete computation of the
thread lowering config until it's needed. This is useful for
distribution of ops like copies and elementwise ops that might first
need to fuse into other operations, so making distribution decisions is
easier after the first level of fusion.