[Codegen][GPU] Add pass to annotate memory spaces on allocations
Trying to infer the memory space of an allocation from within the
bufferization alloc callback function is too late. This adds a
rudimentary pass to annotate the memory space in obvious situations
and then disallows all cases of a bufferization allocation without an
already pre-determined memory space (for the LLVMGPUTileAndFuse
pipeline). This gives us correctness guarantees that were somewhat hand
wavy before.

This pass is still a hack, as all information about the "location" of
a tensor are available during/immediately after tiling. A future TODO is
to further robustify this by annotating memory spaces earlier on. For
now this just disables the on-the-fly inference which should be good
enough given that spurious allocations are almost always the sign of
a performance bug anyway.