[Codegen][GPU] Add support for bufferizing iree_gpu.barrier_region
This adds direct support for bufferizing iree_gpu.barrier_region. Now we
can directly handle this operation during bufferization rather than
requiring it to be decomposed before lowering.

For now this simply bufferizes to two barriers at the beginning and the
end of the region. In the future we could opt to either keep the region,
allowing for some additional analysis, or drop the barriers in certain
cases. Those options are left as TODO and this is kept simple for now.